avin_data_py/pyproject.toml
[project]
name = "avin_data"
version = "0.3.6"
description = "CLI utility for download market data"
authors = [{name = "Alex Avin", email = "mr.alexavin@gmail.com"}]
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.13"
dependencies = [
    "click>=8.1",
    "flit",
    "httpx",
    "moexalgo==2.2.2",
    "numpy",
    "pandas",
    "polars-lts-cpu",
    "pyarrow",
    "pytest",
    "pytest-asyncio",
    "tinkoff-investments",
]

[project.scripts]
avin-data = "avin_data.cli:cli"

[build-system]
requires = ["flit_core >=3.11,<4"]
build-backend = "flit_core.buildapi"

